Add support for objects and array specification on LambdaDslObject
Use of those methods no longer generate invalid Pact body.
The naming of the functions, although terrible, has been kept unchanged in
order to help transitioning from the original API the Lambda one.

There seems to have an inconsistency between `eachArrayWith*Like` and
`*ArrayLike` on how the examples are generated (or rather *not* generated
in the former case). This need to be further investigated as it seems
there is no formal test case for this in the Pact Specifications.