Fix off-by-one valueFromProviderState index
`PactDslJsonArray.valueFromProviderState()` assigns an index
one less than it should. Before appending to the `body` array,
the method invokes `appendArrayIndex`, which in turn emits an array
index value of `body.length()-1`. Since the `body` array is yet to
be updated the calculated index is off by one (the most obvious
sign of this is a negative index of `-1`).

This PR moves the append to `body` first.

(cherry picked from commit 7ab70a91ac97fdffc6c5d6509a595206c92955bb)