feat: Added ignore-order matching for JSON array elements
Adds an EqualsIgnoreOrderMatcher which will uniquely match each item of an actual
array against the expected array, regardless of the order of the actual array items.
For example, [2,3,1] would match [1,2,3]. The EqualsIgnoreOrderMatcher defaults
to Equality matching for non-arrays, such as the items of the array. Matching rules
can also be defined for specific child items of the array using array indexes in the
path (corresponding to the expected item indexes) or for all child items of the array
using a wildcard in the path. EqualsIgnoreOrderMatcher expects the actual array to
be the same size as the expected array.

Adds Min/Max/MinMaxEqualsIgnoreOrderMatcher for cases where array sizes can differ.
Any extra items in the array will only be matched if a wildcard matcher is defined
for the array items.